backend/requirements.txt
# Lumiqe Backend — Python Dependencies
# Install: pip install -r requirements.txt

# Core ML
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Email validation (required by Pydantic EmailStr)
email-validator>=2.0.0
# Computer Vision
opencv-python-headless>=4.8.0
mediapipe>=0.10.0

# API Server (for FastAPI integration)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Image Generation (palette card)
Pillow>=10.0.0

# Auth (JWT validation)
PyJWT>=2.8.0

# Database (PostgreSQL + pgvector)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.3.0
alembic>=1.13.0
python-dotenv>=1.0.0

# Configuration
pydantic-settings>=2.0.0

# Live Scraping (Shopping Agent)
firecrawl-py>=1.0.0

# Security — explicit lower bounds to avoid CVEs in transitive deps
# aiohttp: CVE-2025-69223, CVE-2026-21860, CVE-2026-27199
aiohttp>=3.11.18
# werkzeug: CVE-2025-66221
werkzeug>=3.1.0
# Redis async client (for production rate limiting)
redis[asyncio]>=5.0.0

# Payments
stripe>=8.0.0

# Auth utilities (JWT + password hashing)
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
